==> analysis/02.12_COX.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# Instalar y cargar paquetes necesarios
library(survival)
library(survminer)
library(dplyr)



estudio_vacunas_t_final_selected <- 
  readRDS("/opt/datos_compartidos/vacunas_data/Rfiles/estudio_vacunas_t_final_selected_manuscript.rds")

# Hacer subset para prueba. Vamos a ver mayores de 60 años en periodo delta:
# Convertir valores "Sí" a 1 y "No" a 0 en la variable mortality30d usando dplyr
datitos <- estudio_vacunas_t_final_selected %>% 
  mutate(
    mortality30d = if_else(mortality30d == "Sí", 1, 0)
  #   desc_sexo = if_else(desc_sexo == "Mujer", 0, 1),
  #   across(
  #     c(
  #       desc_sexo, asma, cardiopatia_isquemica, cirrosis_hepatica, diabetes, epoc,
  #       hepatopatia_cronica_excepto_cirrosis, hipertension, insuficiencia_cardiaca,
  #       insuficiencia_renal_cronica, vih, cancer_organo_solido, cancer_hematologico
  #     ),
  #     ~ case_when(.x == "Sí" ~ 1, .x == "No" ~ 0, TRUE ~ NA_real_)),
  #   n_vacunas_covid = case_when(
  #     n_vacunas_covid == "Ninguna" ~ 0,
  #     n_vacunas_covid == "Una" ~ 1,
  #     n_vacunas_covid == "Dos" ~ 2,
  #     n_vacunas_covid == "Tres" ~ 3
  #   ),
  #   n_prev_infecc = case_when(
  #     n_prev_infecc == "Ninguna" ~ 0,
  #     n_prev_infecc == "Uno" ~ 1,
  #     n_prev_infecc == "Dos" ~ 2
  #   ),
  #   variant_period = case_when(
  #     variant_period == "Before α" ~ 0,
  #     variant_period == "α" ~ 1,
  #     variant_period == "α/Δ transition" ~ 2,
  #     variant_period == "Δ" ~ 3,
  #     variant_period == "Δ/Ό transition" ~ 4,
  #     variant_period == "Ό" ~ 5
  #   )
  ) %>%
  filter(
    variant_period == "Δ" 
    & age_fct_60years == "60 años o más"
  )





# Crear variables
event <- "mortality30d"  # Variable de evento 
time_to_event <- "ultima_inmun_infecc_dias"  # Variable tiempo al evento (ultima_inmun_infecc_dias)

# Variables a considerar
covariables <- 
  c(
    "desc_sexo",
    "age", "asma", "cancer_colorrectal", "cancer_de_bronquio_y_pulmon",
    "cancer_de_cabeza_y_cuello", "cancer_de_cuello_uterino",
    "cancer_de_estomago", "cancer_de_higado_y_vias_biliares",
    "cancer_de_hueso_y_tejidos_blandos", "cancer_de_mama",
    "cancer_de_ovario", "cancer_de_pancreas", "cancer_de_prostata",
    "cancer_de_rinon_y_pelvis_renal", "cancer_de_testiculo",
    "cancer_de_tiroides", "cancer_de_utero", "cancer_de_vejiga",
    "cancer_inmunoproliferativo", "cardiopatia_isquemica",
    "cirrosis_hepatica", "diabetes", "enfermedad_de_hodgkin",
    "epoc", "hepatopatia_cronica_excepto_cirrosis", "hipertension",
    "insuficiencia_cardiaca", "insuficiencia_renal_cronica",
    "leucemia", "linfoma_no_hodgkin", "melanoma_de_piel", "sarcoma_de_kaposi",
    "vih",  "n_vacunas_covid", "n_prev_infecc",
    "cancer_organo_solido",
    "cancer_hematologico", "variant_period"
  )




# Ajustar el modelo de supervivencia
cox_fit <- 
  coxph(
    Surv(ultima_inmun_infecc_dias, mortality30d) ~ desc_sexo + age +
      # cancer_colorrectal + cancer_de_bronquio_y_pulmon +
      # cancer_de_cabeza_y_cuello + cancer_de_cuello_uterino +
      # cancer_de_estomago + cancer_de_higado_y_vias_biliares +
      # cancer_de_hueso_y_tejidos_blandos + cancer_de_mama +
      # cancer_de_ovario + cancer_de_pancreas + cancer_de_prostata +
      # cancer_de_rinon_y_pelvis_renal + cancer_de_testiculo +
      # cancer_de_tiroides + cancer_de_utero + cancer_de_vejiga +
      # cancer_inmunoproliferativo + enfermedad_de_hodgkin +
      # leucemia + linfoma_no_hodgkin + melanoma_de_piel + sarcoma_de_kaposi +
      asma + cardiopatia_isquemica + cirrosis_hepatica + diabetes + epoc + 
      hepatopatia_cronica_excepto_cirrosis + hipertension + insuficiencia_cardiaca + 
      insuficiencia_renal_cronica + vih +
      cancer_organo_solido + cancer_hematologico +
      n_vacunas_covid + n_prev_infecc,
      # variant_period,
    data = datitos
  )

# Mostrar el resumen del modelo
summary(cox_fit)

# Crear el gráfico de supervivencia estratificado por grupos
surv_fit <- survfit(cox_fit, newdata = datitos)

plot(surv_fit)

# Create a survival plot using ggplot2 and survminer
surv_plot <- ggsurvplot(surv_fit, 
                        data = datitos,
                        title = "Survival Plot",
                        xlab = "Time",
                        ylab = "Survival Probability")

# Print the survival plot
print(surv_plot)


plot(surv_fit, col = 1:length(variant_period), fun = "event")
legend("topright", legend = levels(factor(variant_period)), col = 1:length(variant_period), lty = 1)



```
