==> analysis/03_tables.Rmd <==
---
title: "03_tables_infections"
# output: html_document
---

# Environment

```{r}
library(tidyverse)
library(lubridate)
library(stringi)
library(janitor)
library(rio)
library(summarytools)
library(zoo)
library(tableone)
library(kableExtra)
library(gtsummary)
library(vroom)
library(parallel)



estudio_general <- 
  readRDS("/opt/datos_compartidos/vacunas_data/Rfiles/estudio_vacunas_t_final_selected_manuscript.rds") %>% 
  select(
    hosp_stay_bin, icu_stay_bin, mortality30d, age, estancia_hospitalaria_dias, num_dias_uci,
    desc_sexo, asma, epoc, hipertension, cardiopatia_isquemica, insuficiencia_cardiaca, diabetes,
    cirrosis_hepatica, hepatopatia_cronica_excepto_cirrosis, insuficiencia_renal_cronica, vih,
    cancer_organo_solido, cancer_hematologico,
    n_vacunas_covid, n_prev_infecc, ultima_vac_infecc_bin, ultima_inmun_infecc_bin2, variant_period
  )

estudio_uci <- 
  estudio_general %>% 
  filter(hosp_stay_bin == "Sí")

```

# Table 1, general data (no strata)

```{r}

table1_general <- CreateTableOne(data = estudio_general)

df_table1_general <-
  print(
    table1_general,
    nonnormal = c("estancia_hospitalaria_dias", "num_dias_uci"),
    quote = FALSE
  ) %>% 
  as_tibble(rownames = "variables") %>% 
  rename(total = Overall) %>% 
  mutate(
    variables = if_else(row_number() %in% c(22:25), paste0("n_vacunas_covid", variables), variables),
    variables = if_else(row_number() %in% c(27:29), paste0("n_prev_infecc", variables), variables),
    variables = if_else(row_number() %in% c(31:35), paste0("ultima_vac_infecc_bin", variables), variables),
    variables = if_else(row_number() %in% c(37:41), paste0("ultima_inmun_infecc_bin", variables), variables),
    variables = if_else(row_number() %in% c(43:48), paste0("variant_period", variables), variables),
  )

```

# Table 1, stratified data: hospital admission

```{r}

table1_hosp <- CreateTableOne(data = estudio_general, strata = "hosp_stay_bin")

df_table1_hosp <-
  print(
    table1_hosp,
    nonnormal = c("estancia_hospitalaria_dias", "num_dias_uci"),
    quote = FALSE
  ) %>% 
  as_tibble(rownames = "variables") %>% 
  select(variables, hospital_no = No, hospital_si = Sí, hospital_p = p) %>% 
  mutate(
    variables = if_else(row_number() %in% c(22:25), paste0("n_vacunas_covid", variables), variables),
    variables = if_else(row_number() %in% c(27:29), paste0("n_prev_infecc", variables), variables),
    variables = if_else(row_number() %in% c(31:35), paste0("ultima_vac_infecc_bin", variables), variables),
    variables = if_else(row_number() %in% c(37:41), paste0("ultima_inmun_infecc_bin", variables), variables),
    variables = if_else(row_number() %in% c(43:48), paste0("variant_period", variables), variables),
  )

```

# Table 1, stratified data: ICU admission

```{r}

table1_icu <- CreateTableOne(data = estudio_uci, strata = "icu_stay_bin")

df_table1_icu <-
  print(
    table1_icu,
    nonnormal = c("estancia_hospitalaria_dias", "num_dias_uci"),
    quote = FALSE
  ) %>% 
  as_tibble(rownames = "variables") %>% 
  select(variables, uci_no = No, uci_si = Sí, uci_p = p) %>% 
  mutate(
    variables = if_else(row_number() %in% c(22:25), paste0("n_vacunas_covid", variables), variables),
    variables = if_else(row_number() %in% c(27:29), paste0("n_prev_infecc", variables), variables),
    variables = if_else(row_number() %in% c(31:35), paste0("ultima_vac_infecc_bin", variables), variables),
    variables = if_else(row_number() %in% c(37:41), paste0("ultima_inmun_infecc_bin", variables), variables),
    variables = if_else(row_number() %in% c(43:48), paste0("variant_period", variables), variables),
  )

```

# Table 1, stratified data: mortality

```{r}

table1_mortality <- CreateTableOne(data = estudio_general, strata = "mortality30d")

df_table1_mortality <-
  print(
    table1_mortality,
    nonnormal = c("estancia_hospitalaria_dias", "num_dias_uci"),
    quote = FALSE
  ) %>% 
  as_tibble(rownames = "variables") %>% 
  select(variables, mortalidad_no = No, mortalidad_si = Sí, mortalidad_p = p) %>% 
  mutate(
    variables = if_else(row_number() %in% c(22:25), paste0("n_vacunas_covid", variables), variables),
    variables = if_else(row_number() %in% c(27:29), paste0("n_prev_infecc", variables), variables),
    variables = if_else(row_number() %in% c(31:35), paste0("ultima_vac_infecc_bin", variables), variables),
    variables = if_else(row_number() %in% c(37:41), paste0("ultima_inmun_infecc_bin", variables), variables),
    variables = if_else(row_number() %in% c(43:48), paste0("variant_period", variables), variables),
  )

```

# Table 1 (altogether)

```{r}

table1 <- 
  left_join(df_table1_general, df_table1_hosp, by = "variables") %>% 
  left_join(df_table1_icu, by = "variables") %>% 
  left_join(df_table1_mortality, by = "variables")

# write.csv(table1, "/opt/datos_compartidos/vacunas_data/tablas/manuscript/altogether.csv")

```



















# Table 1 (deprecated)

```{r}

# str(estudio_vacunas_t_final_selected)



##############################################################################################################
# Description of total data

# t1_total_age <- estudio_vacunas_t_final_selected %>%
#   select(age) %>%
#   tbl_summary(
#     statistic = list(all_continuous() ~ "{mean} ({sd})"),
#     digits = list(age ~ 2)
#   )
# t1_total_age

# t1_total_categ <- estudio_vacunas_t_final_selected %>%
#   select(where(is.factor)) %>%
#   select(-c(mortality, mortality30d, hosp_stay_bin, icu_stay_bin)) %>% 
#   tbl_summary(statistic = list(all_categorical() ~ "{n} ({p}%)"))
# t1_total_categ



##############################################################################################################
# Description by hospital stay

# t1_byhosp_age_hosp_duration <- estudio_vacunas_t_final_selected %>%
#   select(hosp_stay_bin, age, estancia_hospitalaria_dias) %>%
#   tbl_summary(
#     by = hosp_stay_bin,
#     statistic = list(
#       age ~ "{mean} ({sd})",
#       estancia_hospitalaria_dias ~ "{median} ({p25}, {p75})"
#     ),
#     digits = list(age ~ 2)
#   )
#   # add_p() %>%
#   # add_ci()
# t1_byhosp_age_hosp_duration


# t1_byhosp_categ <- estudio_vacunas_t_final_selected %>%
#   select(where(is.factor)) %>%
#   select(-c(mortality, mortality30d, icu_stay_bin)) %>%
#   tbl_summary(
#     by = hosp_stay_bin,
#     percent = "row",
#     statistic = list(all_categorical() ~ "{n} ({p}%)")
#     # digits = list(age ~ 2)
#   )
#   # add_p() %>%
#   # add_ci()
# t1_byhosp_categ



##############################################################################################################
# Description by ICU stay

# t1_byicu_age_icu_duration <- estudio_vacunas_t_final_selected %>%
#   filter(hosp_stay_bin == "Sí") %>% 
#   select(icu_stay_bin, age, estancia_hospitalaria_dias, num_dias_uci) %>%
#   tbl_summary(
#     by = icu_stay_bin,
#     statistic = list(
#       age ~ "{mean} ({sd})",
#       estancia_hospitalaria_dias ~ "{median} ({p25}, {p75})",
#       num_dias_uci ~ "{median} ({p25}, {p75})"
#     ),
#     digits = list(age ~ 2)
#   )
#   # add_p() %>% 
#   # add_ci()
# t1_byicu_age_icu_duration


# t1_byicu_categ <- estudio_vacunas_t_final_selected %>%
#   filter(hosp_stay_bin == "Sí") %>% 
#   select(where(is.factor)) %>%
#   select(-c(mortality, mortality30d, hosp_stay_bin)) %>% 
#   tbl_summary(
#     by = icu_stay_bin,
#     percent = "row",
#     statistic = list(all_categorical() ~ "{n} ({p}%)")
#     # digits = list(age ~ 2)
#   )
#   # add_p() %>% 
#   # add_ci()
# t1_byicu_categ



##############################################################################################################
# Description by mortality at 30 days

# t1_bymortality_age_icu_duration <- estudio_vacunas_t_final_selected %>%
#   select(mortality30d, age, estancia_hospitalaria_dias, num_dias_uci) %>%
#   tbl_summary(
#     by = mortality30d,
#     statistic = list(
#       age ~ "{mean} ({sd})",
#       estancia_hospitalaria_dias ~ "{median} ({p25}, {p75})",
#       num_dias_uci ~ "{median} ({p25}, {p75})"
#     ),
#     digits = list(age ~ 2)
#   )
#   # add_p() %>% 
#   # add_ci()
# t1_bymortality_age_icu_duration


# t1_bymortality_categ <- estudio_vacunas_t_final_selected %>%
#   select(where(is.factor)) %>%
#   select(-c(mortality, icu_stay_bin, hosp_stay_bin)) %>% 
#   tbl_summary(
#     by = mortality30d,
#     percent = "row",
#     statistic = list(all_categorical() ~ "{n} ({p}%)")
#     # digits = list(age ~ 2)
#   )
#   # add_p() %>% 
#   # add_ci()
# t1_bymortality_categ

```
